cloud-function/requirements.txt
google-cloud-firestore==2.14.0
google-cloud-pubsub==2.19.0
